# Optimization for Float floor and ceil functions when precision is 0