@michalmuskala 2018-07-12
How does using `:math.floor` and `:math.ceil` compare?