@amatalai 2018-07-12
`:math.ceil/1` and `:math.floor/1` are faster

```
Name                         ips        average  deviation         median         99th %
erlang ceil             23214.94       43.08 μs    ±17.66%          42 μs          81 μs
math ceil               39032.17       25.62 μs    ±30.36%          25 μs          48 μs

erlang floor            23373.19       42.78 μs    ±15.30%          42 μs          80 μs
math floor              46143.92       21.67 μs    ±21.94%          21 μs          39 μs
```

Also, according to benchee `:math` version uses less memory

thank you @michalmuskala