@amatalai 2018-07-12
My first commit doesn't make much sense now. Should I rebase them and force push?